Renamed Hangups System to Core, moved Auth to Core
Auth has a single public method called "PerformOauth" that handles
whether or not to show the oauth broker to the user. all other methods
are called behind the scenes. Also, added an error message to the
notification helper. This makes it so errors in the auth layer(or
anywhere) can be handled with a message.